Use short instead of int or unsigned where it makes sense.
This saves some storage space.

+++ src/A4988.cpp
@@ -19,11 +19,11 @@ const uint8_t A4988::MS_TABLE[] = {0b000, 0b001, 0b010, 0b011, 0b111};
  * Basic connection: only DIR, STEP are connected.
  * Microstepping controls should be hardwired.
  */
-A4988::A4988(int steps, int dir_pin, int step_pin)
+A4988::A4988(short steps, short dir_pin, short step_pin)
 :BasicStepperDriver(steps, dir_pin, step_pin)
 {}
 
-A4988::A4988(int steps, int dir_pin, int step_pin, int enable_pin)
+A4988::A4988(short steps, short dir_pin, short step_pin, short enable_pin)
 :BasicStepperDriver(steps, dir_pin, step_pin, enable_pin)
 {}
 
@@ -31,17 +31,17 @@ A4988::A4988(int steps, int dir_pin, int step_pin, int enable_pin)
  * Fully wired.
  * All the necessary control pins for A4988 are connected.
  */
-A4988::A4988(int steps, int dir_pin, int step_pin, int ms1_pin, int ms2_pin, int ms3_pin)
+A4988::A4988(short steps, short dir_pin, short step_pin, short ms1_pin, short ms2_pin, short ms3_pin)
 :BasicStepperDriver(steps, dir_pin, step_pin),
     ms1_pin(ms1_pin), ms2_pin(ms2_pin), ms3_pin(ms3_pin)
 {}
 
-A4988::A4988(int steps, int dir_pin, int step_pin, int enable_pin, int ms1_pin, int ms2_pin, int ms3_pin)
+A4988::A4988(short steps, short dir_pin, short step_pin, short enable_pin, short ms1_pin, short ms2_pin, short ms3_pin)
 :BasicStepperDriver(steps, dir_pin, step_pin, enable_pin),
 ms1_pin(ms1_pin), ms2_pin(ms2_pin), ms3_pin(ms3_pin)
 {}
 
-void A4988::begin(int rpm, unsigned microsteps){
+void A4988::begin(short rpm, short microsteps){
     BasicStepperDriver::begin(rpm, microsteps);
 
     if (!IS_CONNECTED(ms1_pin) || !IS_CONNECTED(ms2_pin) || !IS_CONNECTED(ms3_pin)){
@@ -58,7 +58,7 @@ void A4988::begin(int rpm, unsigned microsteps){
  * Allowed ranges for A4988 are 1:1 to 1:16
  * If the control pins are not connected, we recalculate the timing only
  */
-unsigned A4988::setMicrostep(unsigned microsteps){
+short A4988::setMicrostep(short microsteps){
     BasicStepperDriver::setMicrostep(microsteps);
 
     if (!IS_CONNECTED(ms1_pin) || !IS_CONNECTED(ms2_pin) || !IS_CONNECTED(ms3_pin)){
@@ -68,7 +68,7 @@ unsigned A4988::setMicrostep(unsigned microsteps){
     const uint8_t* ms_table = getMicrostepTable();
     size_t ms_table_size = getMicrostepTableSize();
 
-    unsigned i = 0;
+    unsigned short i = 0;
     while (i < ms_table_size){
         if (this->microsteps & (1<<i)){
             uint8_t mask = ms_table[i];
@@ -90,6 +90,6 @@ size_t A4988::getMicrostepTableSize(){
     return sizeof(A4988::MS_TABLE);
 }
 
-unsigned A4988::getMaxMicrostep(){
+short A4988::getMaxMicrostep(){
     return A4988::MAX_MICROSTEP;
 }
